Move Results and Scrambles code to shared folder
