--- app/views/competitions/_scrambles_nav.html.erb
@@ -1,18 +1,0 @@
-<% content = capture do %>
-  <%= render "competition_info", competition: @competition %>
-
-  <% if @competition.user_can_view_results?(current_user) %>
-    <% if !@competition.results_posted? %>
-      <%= alert :warning, t('competitions.messages.results_preview_alert'), note: false %>
-    <% end %>
-    <%= yield %>
-  <% end %>
-<% end %>
-
-<% if !defined?(layout_nav) || layout_nav %>
-  <%= render layout: 'nav' do %>
-    <%= content %>
-  <% end %>
-<% else %>
-  <%= content %>
-<% end %>
