+++ app/views/competitions/show_all_results.html.erb
@@ -2,7 +2,7 @@
 
 <%= render layout: 'results_nav' do %>
   <%= tag :div, id: "selected-event", "data-selected": params[:event] %>
-  <%= react_component("CompetitionResults", {
+  <%= react_component("ResultsData/Results", {
     competitionId: @competition.id,
     canAdminResults: current_user&.can_admin_results?,
   }, {
