+++ app/webpacker/components/ResultsData/EventNavigation.js
@@ -1,9 +1,9 @@
 import classnames from 'classnames';
 import { Popup, Menu } from 'semantic-ui-react';
 import React from 'react';
-import EventIcon from './wca/EventIcon';
-import { events } from '../lib/wca-data.js.erb';
-import '../stylesheets/event_navigation.scss';
+import EventIcon from '../wca/EventIcon';
+import { events } from '../../lib/wca-data.js.erb';
+import '../../stylesheets/event_navigation.scss';
 
 function EventNavigation({ selected, eventIds, onSelect }) {
   return (
